git-svn-id: https://svn.dealii.org/branches/branch_cmake@27186 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/include/deal.II/lac/trilinos_vector.h
@@ -537,7 +537,7 @@ namespace TrilinosWrappers
       if (size() != v.size())
         {
           vector.reset (new Epetra_FEVector(Epetra_Map
-                                            (v.size(), 0,
+                                            (static_cast<int>(v.size()), 0,
 #ifdef DEAL_II_COMPILER_SUPPORTS_MPI
                                              Epetra_MpiComm(MPI_COMM_SELF)
 #else
